Attempt to fix aliases and TS server red underline
