+++ package.json
@@ -7,7 +7,7 @@
     "dev": "vite",
     "build": "vue-tsc && vite build",
     "preview": "vite preview",
-    "tests": "vitest tests/ --"
+    "tests": "vitest tests/"
   },
   "dependencies": {
     "@fluentui/web-components": "^2.5.12",
